Aqua.jl

Aqua.Aqua — Module

Aqua.jl: Auto QUality Assurance for Julia packages

Aqua.jl provides functions to run a few automatable checks for Julia packages:

There are no method ambiguities.

There are no undefined exports.

There are no unbound type parameters.

There are no stale dependencies listed in Project.toml.

Check that test target of the root project Project.toml and test project (test/Project.toml) are consistent.

Check that all external packages listed in deps have corresponding compat entry.

Project.toml formatting is compatible with Pkg.jl output.

There are no "obvious" type piracies (new in 0.6)

See more in the documentation.

Quick usage

Call Aqua.test_all(YourPackage) from test/runtests.jl, e.g.,

using YourPackage
using Aqua
Aqua.test_all(YourPackage)

Notes on Aqua 0.6

Aqua 0.6 includes the type piracy detection, thanks to the PR by Jakob Nybo Nissen (@jakobnissen) and the original implementation by Frames Catherine White (@oxinabox).

If this part of Aqua 0.6 causes a trouble, then you can disable the piracy detection with a flag as in Aqua.test_all(YourPackage; piracy = false).

Specifying Aqua version

To avoid breaking test when a new Aqua.jl version is released, it is recommended to add version bound for Aqua.jl in test/Project.toml:

[deps]
Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"

[compat]
Aqua = "0.6"

Badge

You can add the following line in README.md to include Aqua.jl badge:

[![Aqua QA](https://raw.githubusercontent.com/JuliaTesting/Aqua.jl/master/badge.svg)](https://github.com/JuliaTesting/Aqua.jl)

which is rendered as

Aqua.test_all — Method

test_all(testtarget::Module)

Run following tests in isolated testset:

test_ambiguities([testtarget, Base, Core])

test_unbound_args(testtarget)

test_undefined_exports(testtarget)

test_project_extras(testtarget)

test_stale_deps(testtarget)

test_deps_compat(testtarget)

test_project_toml_formatting(testtarget)

test_piracy(testtarget)

Aqua.jl 0.5

Since Aqua.jl 0.5:

test_all runs test_ambiguities with Core. This means method ambiguities of constructors can now be detected. In Aqua.jl 0.4, test_ambiguities was invoked with [testtarget, Base].

test_all runs test_stale_deps. In Aqua.jl 0.4, this check was opt-in.

The keyword argument $x (e.g., ambiguities) can be used to control whether or not to run test_$x (e.g., test_ambiguities). If test_$x supports keyword arguments, a NamedTuple can also be passed to $x to specify the keyword arguments for test_$x.

Keyword Arguments

ambiguities = true

unbound_args = true

undefined_exports = true

project_extras = true

stale_deps = true

deps_compat = true

project_toml_formatting = true

piracy = true

Aqua.test_ambiguities — Method

test_ambiguities(package::Union{Module, PkgId})
test_ambiguities(packages::Vector{Union{Module, PkgId}})

Test that there is no method ambiguities in given package(s). It calls Test.detect_ambiguities in a separated clean process to avoid false-positive.

Keyword Arguments

broken::Bool = false: If true, it uses @test_broken instead of @test.

color::Union{Bool, Nothing} = nothing: Enable/disable colorful output if a Bool. nothing (default) means to inherit the setting in the current process.

exclude::AbstractVector = []: A vector of functions or types to be excluded from ambiguity testing. A function means to exclude all its methods. A type means to exclude all its methods of the callable (sometimes also called "functor"). That is to say, MyModule.MyType means to ignore ambiguities between (::MyType)(x, y::Int) and (::MyType)(x::Int, y). Note that there is no way to exclude the constructor of a specific type at the moment.

recursive::Bool = true: Passed to Test.detect_ambiguities. Note that the default here (true) is different from detect_ambiguities. This is for testing ambiguities in methods defined in all sub-modules.

Other keyword arguments such as imported and ambiguous_bottom are passed to Test.detect_ambiguities as-is.

Aqua.test_deps_compat — Method

Aqua.test_deps_compat(package; [ignore])

Test that Project.toml of package list all compat for deps.

Arguments

packages: a top-level Module, a Base.PkgId, or a collection of them.

Keyword Arguments

ignore::Vector{Symbol}: names of dependent packages to be ignored.

Aqua.test_piracy — Method

test_piracy(m::Module)

Test that m does not commit type piracy. See Julia documentation for more information about type piracy.

Keyword Arguments

broken::Bool = false: If true, it uses @test_broken instead of @test.

treat_as_own = Union{Function, Type}[]: The types in this container are considered to be "owned" by the module m. This is useful for testing packages that deliberately commit some type piracy, e.g. modules adding higher-level functionality to a lightweight C-wrapper, or packages that are extending StatsAPI.jl.

Aqua.test_project_extras — Method

test_project_extras(package::Union{Module, PkgId})
test_project_extras(packages::Vector{Union{Module, PkgId}})

Check that test target of the root project and test project (test/Project.toml) are consistent. This is useful for supporting Julia < 1.2 while recording test-only dependency compatibility in test/Project.toml.

Aqua.test_project_toml_formatting — Method

Aqua.test_project_toml_formatting(packages)

Aqua.test_stale_deps — Method

Aqua.test_stale_deps(package; [ignore])

Test that package loads all dependencies listed in Project.toml. Note that this does not imply that package loads the dependencies directly, this can be achieved via transitivity as well.

Known bug

Currently, Aqua.test_stale_deps does not detect stale dependencies when they are stdlib. This is considered a bug and may be fixed in the future. Such a release is considered non-breaking.

Arguments

packages: a top-level Module, a Base.PkgId, or a collection of them.

Keyword Arguments

ignore::Vector{Symbol}: names of dependent packages to be ignored.

Aqua.test_unbound_args — Method

test_unbound_args(module::Module)

Test that all methods in module and its submodules do not have unbound type parameters. An unbound type parameter is a type parameter with a where, that does not occurr in the signature of some dispatch of the method.

Keyword Arguments

broken::Bool = false: If true, it uses @test_broken instead of @test.

For example, the following methods each have T as an unbound type parameter:

f(x::Int) where {T} = do_something(x)

g(x::S) where {S < : Number, T <: Number} = do_something(x)

h(x::T...) where {T} = do_something(x)

In the cases of f and g above, the unbound type parameter T is neither present in the signature of the methods nor as a bound of another type parameter. Here, the type parameter T can be removed without changing any semantics.

For signatures with Vararg (cf. h above), the type parameter unbound for the zero-argument case (e.g. h()). A possible fix would be to replace h by two methods

h() = do_something(T[])
h(x1::T, x2::T...) = do_something(T[x1, x2...])

Aqua.test_undefined_exports — Method

test_undefined_exports(module::Module)

Test that all exported names in module actually exist.

Keyword Arguments

broken::Bool = false: If true, it uses @test_broken instead of @test.