Compat entries

In your Project.toml you can (and should) use compat entries to specify with which versions of Julia and your dependencies your package is compatible with. This is important to ease the installation and upgrade of your package for users, and to keep everything working in the case of breaking changes in Julia or your dependencies.

For more details, see the Pkg docs.

Test function

Aqua.test_deps_compat — Function

Aqua.test_deps_compat(package)

Test that the Project.toml of package has a compat entry for each package listed under deps and for julia.

Arguments

packages: a top-level Module, a Base.PkgId, or a collection of them.

Keyword Arguments

Test choosers

check_julia = true: If true, additionally check for a compat entry for "julia".

check_extras = true: If true, additionally check "extras". A NamedTuple can be used to pass keyword arguments with test options (see below).

check_weakdeps = true: If true, additionally check "weakdeps". A NamedTuple can be used to pass keyword arguments with test options (see below).

Test options

If these keyword arguments are set directly, they only apply to the standard test for "deps". To apply them to "extras" and "weakdeps", pass them as a NamedTuple to the corresponding check_$x keyword argument.

broken::Bool = false: If true, it uses @test_broken instead of @test for "deps".

ignore::Vector{Symbol}: names of dependent packages to be ignored.