Stale dependencies

Test function

Aqua.test_stale_deps — Function

Aqua.test_stale_deps(package; [ignore])

Test that package loads all dependencies listed in Project.toml. Note that this does not imply that package loads the dependencies directly, this can be achieved via transitivity as well.

Weak dependencies and extensions

Due to the automatic loading of package extensions once all of their trigger dependencies are loaded, Aqua.jl can, by design of julia, not check if a package extension indeed loads all of its trigger dependencies using import or using.

Known bug

Currently, Aqua.test_stale_deps does not detect stale dependencies when they are in the sysimage. This is considered a bug and may be fixed in the future. Such a release is considered non-breaking.

Arguments

packages: a top-level Module, a Base.PkgId, or a collection of them.

Keyword Arguments

ignore::Vector{Symbol}: names of dependent packages to be ignored.