Test everything

This test runs most of the other tests in this module. The defaults should be fine for most packages. If you have a package that needs to customize the test, you can do so by providing appropriate keyword arguments to Aqua.test_all() (see below)

Aqua.test_all — Function

test_all(testtarget::Module)

Run the following tests on the module testtarget:

test_ambiguities([testtarget])

test_unbound_args(testtarget)

test_undefined_exports(testtarget)

test_project_extras(testtarget)

test_stale_deps(testtarget)

test_deps_compat(testtarget)

test_piracies(testtarget)

test_persistent_tasks(testtarget)

test_undocumented_names(testtarget)

The keyword argument $x (e.g., ambiguities) can be used to control whether or not to run test_$x (e.g., test_ambiguities). If test_$x supports keyword arguments, a NamedTuple can also be passed to $x to specify the keyword arguments for test_$x.

Keyword Arguments

ambiguities = true

unbound_args = true

undefined_exports = true

project_extras = true

stale_deps = true

deps_compat = true

piracies = true

persistent_tasks = true

undocumented_names = false