Undocumented names

Test function

Aqua.test_undocumented_names — Function

test_undocumented_names(module::Module)

Test that all public names in module and its recursive submodules have a docstring (not including module itself).

Tip

On all Julia versions, public names include the exported names. On Julia versions >= 1.11, public names also include the names annotated with the public keyword.

Warning

When running this Aqua test in Julia versions before 1.11, it does nothing. Thus if you use continuous integration tests, make sure those are configured to use Julia >= 1.11 in order to benefit from this test.

Keyword Arguments

broken::Bool = false: If true, it uses @test_broken instead of @test and shortens the error message.