Aqua.jl:

Auto QUality Assurance for Julia packages

Aqua.jl provides functions to run a few automatable checks for Julia packages:

There are no method ambiguities.

There are no undefined exports.

There are no unbound type parameters.

There are no stale dependencies listed in Project.toml.

Check that test target of the root project Project.toml and test project (test/Project.toml) are consistent.

Check that all external packages listed in deps have corresponding compat entries.

There are no "obvious" type piracies.

The package does not create any persistent Tasks that might block precompilation of dependencies.

Quick usage

Call Aqua.test_all(YourPackage) from the REPL, e.g.,

using YourPackage
using Aqua
Aqua.test_all(YourPackage)

How to add Aqua.jl...

...as a test dependency?

There are two ways to add Aqua.jl as a test dependency to your package. To avoid breaking tests when a new Aqua.jl version is released, it is recommended to add a version bound for Aqua.jl.

In YourPackage/test/Project.toml, add Aqua.jl to [dep] and [compat] sections, like

[deps]
Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"

[compat]
Aqua = "0.8"

In YourPackage/Project.toml, add Aqua.jl to [compat] and [extras] section and the test target, like

[compat]
Aqua = "0.8"

[extras]
Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"

[targets]
test = ["Aqua", "Test"]

If your package supports Julia pre-1.2, you need to use the second approach, although you can use both approaches at the same time.

Warning

In normal use, Aqua.jl should not be added to [deps] in YourPackage/Project.toml!

...to your tests?

It is recommended to create a separate file YourPackage/test/Aqua.jl that gets included in YourPackage/test/runtests.jl with either

using Aqua
Aqua.test_all(YourPackage)

or some fine-grained checks with options, e.g.,

using Aqua

@testset "Aqua.jl" begin
  Aqua.test_all(
    YourPackage;
    ambiguities=(exclude=[SomePackage.some_function], broken=true),
    stale_deps=(ignore=[:SomePackage],),
    deps_compat=(ignore=[:SomeOtherPackage],),
    piracies=false,
  )
end

Note, that for all tests with no explicit options provided, the default options are used.

For more details on the options, see the respective functions below.

Example uses

The following is a small selection of packages that use Aqua.jl:

GAP.jl

Hecke.jl

Oscar.jl

Test functions

Aqua.test_all — Function

test_all(testtarget::Module)

Run following tests in isolated testset:

test_ambiguities([testtarget, Base, Core])

test_unbound_args(testtarget)

test_undefined_exports(testtarget)

test_project_extras(testtarget)

test_stale_deps(testtarget)

test_deps_compat(testtarget)

test_piracies(testtarget)

test_persistent_tasks(testtarget)

The keyword argument $x (e.g., ambiguities) can be used to control whether or not to run test_$x (e.g., test_ambiguities). If test_$x supports keyword arguments, a NamedTuple can also be passed to $x to specify the keyword arguments for test_$x.

Keyword Arguments

ambiguities = true

unbound_args = true

undefined_exports = true

project_extras = true

stale_deps = true

deps_compat = true

piracies = true

persistent_tasks = true

Aqua.find_persistent_tasks_deps — Method

Aqua.find_persistent_tasks_deps(package; broken = Dict{String,Bool}(), kwargs...)

Test all the dependencies of package with Aqua.test_persistent_tasks. On Julia 1.10 and higher, it returns a list of all dependencies failing the test. These are likely the ones blocking precompilation of your package.

Any additional kwargs (e.g., tmax) are passed to Aqua.test_persistent_tasks.

Aqua.test_ambiguities — Method

test_ambiguities(package::Union{Module, PkgId})
test_ambiguities(packages::Vector{Union{Module, PkgId}})

Test that there is no method ambiguities in given package(s). It calls Test.detect_ambiguities in a separated clean process to avoid false-positive.

Keyword Arguments

broken::Bool = false: If true, it uses @test_broken instead of @test.

color::Union{Bool, Nothing} = nothing: Enable/disable colorful output if a Bool. nothing (default) means to inherit the setting in the current process.

exclude::AbstractVector = []: A vector of functions or types to be excluded from ambiguity testing. A function means to exclude all its methods. A type means to exclude all its methods of the callable (sometimes also called "functor") and the constructor. That is to say, MyModule.MyType means to ignore ambiguities between (::MyType)(x, y::Int) and (::MyType)(x::Int, y).

recursive::Bool = true: Passed to Test.detect_ambiguities. Note that the default here (true) is different from detect_ambiguities. This is for testing ambiguities in methods defined in all sub-modules.

Other keyword arguments such as imported and ambiguous_bottom are passed to Test.detect_ambiguities as-is.

Aqua.test_deps_compat — Method

Aqua.test_deps_compat(package)

Test that the Project.toml of package has a compat entry for each package listed under deps and for julia.

Arguments

packages: a top-level Module, a Base.PkgId, or a collection of them.

Keyword Arguments

Test choosers

check_julia = true: If true, additionally check for a compat entry for "julia".

check_extras = true: If true, additionally check "extras". A NamedTuple can be used to pass keyword arguments with test options (see below).

check_weakdeps = true: If true, additionally check "weakdeps". A NamedTuple can be used to pass keyword arguments with test options (see below).

Test options

If these keyword arguments are set directly, they only apply to the standard test for "deps". To apply them to "extras" and "weakdeps", pass them as a NamedTuple to the corresponding check_$x keyword argument.

broken::Bool = false: If true, it uses @test_broken instead of @test for "deps".

ignore::Vector{Symbol}: names of dependent packages to be ignored.

Aqua.test_persistent_tasks — Method

Aqua.test_persistent_tasks(package)

Test whether loading package creates persistent Tasks which may block precompilation of dependent packages. See also Aqua.find_persistent_tasks_deps.

Motivation

Julia 1.10 and higher wait for all running Tasks to finish before writing out the precompiled (cached) version of the package. One consequence is that a package that launches Tasks in its __init__ function may precompile successfully, but block precompilation of any packages that depend on it.

Example

Let's create a dummy package, PkgA, that launches a persistent Task:

module PkgA
const t = Ref{Any}()   # to prevent the Timer from being garbage-collected
__init__() = t[] = Timer(0.1; interval=1)   # create a persistent `Timer` `Task`
end

PkgA will precompile successfully, because PkgA.__init__() does not run when PkgA is precompiled. However,

module PkgB
using PkgA
end

fails to precompile: using PkgA runs PkgA.__init__(), which leaves the Timer Task running, and that causes precompilation of PkgB to hang.

How the test works

This test works by launching a Julia process that tries to precompile a dummy package similar to PkgB above, modified to signal back to Aqua when PkgA has finished loading. The test fails if the gap between loading PkgA and finishing precompilation exceeds time tmax.

How to fix failing packages

Often, the easiest fix is to modify the __init__ function to check whether the Julia process is precompiling some other package; if so, don't launch the persistent Tasks.

function __init__()
    # Other setup code here
    if ccall(:jl_generating_output, Cint, ()) == 0   # if we're not precompiling...
        # launch persistent tasks here
    end
end

In more complex cases, you may need to set up independently-callable functions to launch the tasks and set conditions that allow them to cleanly exit.

On Julia version 1.9 and before, this test always succeeds.

Arguments

package: a top-level Module or Base.PkgId.

Keyword Arguments

broken::Bool = false: If true, it uses @test_broken instead of @test.

tmax::Real = 5: the maximum time (in seconds) to wait after loading the package before forcibly shutting down the precompilation process (triggering a test failure).

Aqua.test_piracies — Method

test_piracies(m::Module)

Test that m does not commit type piracies. See Julia documentation for more information about type piracy.

Keyword Arguments

broken::Bool = false: If true, it uses @test_broken instead of @test.

skip_deprecated::Bool = true: If true, it does not check deprecated methods.

treat_as_own = Union{Function, Type}[]: The types in this container are considered to be "owned" by the module m. This is useful for testing packages that deliberately commit some type piracies, e.g. modules adding higher-level functionality to a lightweight C-wrapper, or packages that are extending StatsAPI.jl.

Aqua.test_project_extras — Method

test_project_extras(package::Union{Module, PkgId})
test_project_extras(packages::Vector{Union{Module, PkgId}})

Check that test target of the root project and test project (test/Project.toml) are consistent. This is useful for supporting Julia < 1.2 while recording test-only dependency compatibility in test/Project.toml.

Aqua.test_stale_deps — Method

Aqua.test_stale_deps(package; [ignore])

Test that package loads all dependencies listed in Project.toml. Note that this does not imply that package loads the dependencies directly, this can be achieved via transitivity as well.

Weak dependencies and extensions

Due to the automatic loading of package extensions once all of their trigger dependencies are loaded, Aqua.jl can, by design of julia, not check if a package extension indeed loads all of its trigger dependencies using import or using.

Known bug

Currently, Aqua.test_stale_deps does not detect stale dependencies when they are stdlib. This is considered a bug and may be fixed in the future. Such a release is considered non-breaking.

Arguments

packages: a top-level Module, a Base.PkgId, or a collection of them.

Keyword Arguments

ignore::Vector{Symbol}: names of dependent packages to be ignored.

Aqua.test_unbound_args — Method

test_unbound_args(module::Module)

Test that all methods in module and its submodules do not have unbound type parameters. An unbound type parameter is a type parameter with a where, that does not occur in the signature of some dispatch of the method.

Keyword Arguments

broken::Bool = false: If true, it uses @test_broken instead of @test.

For example, the following methods each have T as an unbound type parameter:

f(x::Int) where {T} = do_something(x)

g(x::S) where {S < : Number, T <: Number} = do_something(x)

h(x::T...) where {T} = do_something(x)

In the cases of f and g above, the unbound type parameter T is neither present in the signature of the methods nor as a bound of another type parameter. Here, the type parameter T can be removed without changing any semantics.

For signatures with Vararg (cf. h above), the type parameter unbound for the zero-argument case (e.g. h()). A possible fix would be to replace h by two methods

h() = do_something(T[])
h(x1::T, x2::T...) = do_something(T[x1, x2...])

Aqua.test_undefined_exports — Method

test_undefined_exports(module::Module)

Test that all exported names in module actually exist.

Keyword Arguments

broken::Bool = false: If true, it uses @test_broken instead of @test.